requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==25.4.0
billiard==4.2.4
black==26.3.1
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
croniter==6.2.2
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.9.4
prompt_toolkit==3.0.52
PyJWT==2.12.0
python-dateutil==2.9.0.post0
pytokens==0.4.1
PyYAML==6.0.3
redis==7.3.0
referencing==0.37.0
rpds-py==0.30.0
ruff==0.15.6
setuptools==78.1.1
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.6.0
wheel==0.45.1
